clean still in progress (grocery list)
