--- lib/screens/home/contextual_area/consts.dart
@@ -1,6 +1,0 @@
-// Keys for recipe of the week
-const KEY_RECIPE_OF_THE_WEEK_LABEL = 0;
-const KEY_RECIPE_OF_THE_WEEK_IMG = 1;
-const KEY_RECIPE_OF_THE_WEEK_URL = 2;
-const KEY_RECIPE_OF_THE_WEEK_CALORIES = 3;
-const KEY_RECIPE_OF_THE_WEEK_COOKING_INGREDIENTS = 4;
